+++ tests/tests_adapters/test_adapter_provider5.py
@@ -1,12 +1,12 @@
 from typing import List
 
-from th2_data_services.adapter import change_pipeline_message
+from th2_data_services.adapters.adapter_provider5 import adapter_provider5
 
 
 def test_adapter_with_id(messages_before_pipeline_adapter: List[dict], messages_after_pipeline_adapter: List[dict]):
     output = []
     for message in messages_before_pipeline_adapter:
-        messages = change_pipeline_message(message)
+        messages = adapter_provider5(message)
         if isinstance(messages, dict):
             messages = [messages]
         output.extend(messages)
@@ -16,7 +16,7 @@ def test_adapter_with_id(messages_before_pipeline_adapter: List[dict], messages_
 
 def test_find_sub_message_with_adapter(message_from_pipeline: dict):
     msg_id = message_from_pipeline.get("messageId") + ".5"
-    message_from_pipeline = change_pipeline_message(message_from_pipeline)
+    message_from_pipeline = adapter_provider5(message_from_pipeline)
 
     index = None
     if msg_id.find(".") != -1:
@@ -73,7 +73,7 @@ def test_find_sub_message_with_adapter(message_from_pipeline: dict):
 
 def test_find_messages_with_adapter(message_from_pipeline: dict):
     msg_id = message_from_pipeline.get("messageId")
-    message_from_pipeline = change_pipeline_message(message_from_pipeline)
+    message_from_pipeline = adapter_provider5(message_from_pipeline)
 
     index = None
     if msg_id.find(".") != -1:
@@ -95,6 +95,6 @@ def test_find_messages_with_adapter(message_from_pipeline: dict):
 
 
 def test_message_with_empty_body(message_from_pipeline_empty_body, messages_from_after_pipeline_empty_body):
-    messages = change_pipeline_message(message_from_pipeline_empty_body)
+    messages = adapter_provider5(message_from_pipeline_empty_body)
 
     assert messages_from_after_pipeline_empty_body == messages