+++ release_notes.md
@@ -116,6 +116,20 @@
 
 # v0.6.0
 
-## Features
+## User impact and migration instructions
 
-1. [TH2-2283] Implemented rpt-data-provider Filters API.
+## Improvements
+
+## Features
+1. You can import classes directly from th2_data_services now.  
+   E.g. `from th2_data_services import Data, DataSource, Filter`
+2. [TH2-2755] Added new parameters to DataSource methods.
+   1. get_X_from_data_provider:
+      - sse_adapter=True. If True, all data will go through SSE adapter and yield dicts.
+                   Otherwise adapter will yield SSE Events. 
+   2. get_messages_from_data_provider and find_messages_by_id_from_data_provider:
+      - provider_adapter=adapter_provider5. Adapter function for rpt-data-provider.
+                   If None, Data object will yield object from previous map function.
+
+
+## BugFixes